Issue #1: Allow archiving of applications
Event administrators and superadmins can now archive applications. This means this applications will not be visible to event leaders.

This is done clicking in a checkbox in the application info.

Practical example: application gathered during the call for prep-team can be archived so that the PT itself cannot see them when they become leaders to the activity.

Superadmins can now change the status of an application status and archive them too.